+++ Resources/Private/Templates/Container.html
@@ -1,0 +1,6 @@
+<html
+    xmlns:f="http://typo3.org/ns/TYPO3/CMS/Fluid/ViewHelpers"
+    data-namespace-typo3-fluid="true"
+>
+<f:format.raw>{tx_container_grid}</f:format.raw>
+</html>
